move code around, fix flash of alert
