+++ client/src/features/game/GameRoutes.tsx
@@ -1,6 +1,6 @@
 import { Route, Switch, useParams } from "react-router-dom";
 
-import { GameProvider } from "features/game/GameChannel";
+import { GameProvider } from "features/game/GameProvider";
 import { GameLobby } from "features/game/GameLobby";
 import { GameSpectate } from "features/game/GameSpectate";
 import { Game } from "features/game/Game";
