+++ client/src/index.tsx
@@ -4,7 +4,7 @@ import { debounce } from "lodash";
 import { ApolloProvider } from "@apollo/client";
 import { Provider as ReduxProvider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
-import { SocketProvider } from "utils/Socket";
+import { SocketProvider } from "utils/SocketProvider";
 
 import { store } from "app/store";
 import { client } from "app/apollo";
